requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.2
tqdm==4.66.5
openai==1.51.2
python-dotenv==1.0.1
scikit-learn==1.5.2
httpx==0.27.2
